read/write fst files functions
this is necesary for pak tool to work

+++ mclib/ffent.h
@@ -14,7 +14,7 @@
 
 #pragma pack(1)
 typedef struct {
-    long	offset;
+    int32_t offset;
     char	name[250];
 } FILEENTRY, *LPFILEENTRY;
 #pragma pack()
